requirements.txt
streamlit==1.32.0
langchain==0.1.11
google-generativeai
langchain_google_genai
pyPDF2==3.0.1
python-dotenv==1.0.1 
faiss-cpu==1.8.0